v4l-utils: fix racy udev rule
ir-keytable -a can fail if it's run before the input and event
devices are created. Fix this by changing the udev rule to trigger
on event device creation instead of rc device creation.

Also drop the udev rules patch and simply replace the udev rule
with our version, this is easier to maintain.

Signed-off-by: Matthias Reichl <hias@horus.com>